ecsevent/requirements.txt
aniso8601==3.0.2
werkzeug==0.16.0
chardet==3.0.2
baidu-aip
ldap3==2.5.1
passlib==1.7.1
gevent==1.3.7
gunicorn==19.9.0
alembic==1.0.0
bcrypt==3.1.4
pytz==2018.5
PyMySQL==0.9.2
PyYAML==3.12
PyJWT==1.6.4
redis==2.10.6
rq==0.12.0
rq_scheduler==0.8.3
Flask==0.12.4
Flask-RQ2==18.2.1
Flask-Migrate==2.3.1
Flask-Debugtoolbar==0.10.1
Flask-SQLAlchemy==2.3.2
Flask-Caching==1.4.0
Flask-Limiter==1.0.1
Flasgger==0.9.2
sqlalchemy==1.2.12
Sqlalchemy-Utils==0.34.1
setproctitle==1.1.10
#ansible==2.5.7
pypinyin==0.33.2
gitpython==2.1.11
python-jenkins==0.4.16
pycryptodome
xmltodict==0.11.0
markdown==3.0.1
BeautifulSoup4==4.6.3
kazoo==2.5.0
xlwt==1.3.0
selenium==3.141.0
beautifulsoup4==4.6.3 
html5lib==1.0.1 
#: Aliyun
aliyun-python-sdk-bssopenapi==1.5.4
aliyun-python-sdk-sts==3.0.0
aliyun-python-sdk-core-v3==2.9.3
aliyun-python-sdk-ecs==4.18.3
aliyun-python-sdk-actiontrail==2.0.0
aliyun-python-sdk-slb==3.2.7
aliyun-python-sdk-rds==2.3.0
aliyun-python-sdk-ram==3.1.0
#: Baidu
#: baidu-aip==2.0.0.1
pyecharts~=1.9.0
snapshot_selenium
pandas~=1.1.5
python-gitlab==2.5.0

python-dateutil~=2.8.1
croniter~=1.0.1
uuid~=1.30
DateTime~=4.3
Click~=7.0
greenlet~=1.0.0
requests~=2.22.0
gitdb~=4.0.5
cryptography~=3.1.1
pbr~=5.5.1
setuptools~=39.0.1
six~=1.14.0
virtualenv~=20.0.29
pip~=9.0.3
py~=1.9.0
attrs~=20.2.0
Mako~=1.1.2
MarkupSafe~=1.1.1
colorama~=0.4.3
Jinja2~=2.10.3
itsdangerous~=1.1.0
blinker~=1.4
smmap~=3.0.4
pyasn1~=0.4.8
pytest~=6.1.0
numpy~=1.19.2
cffi~=1.14.3
limits~=1.5.1
certifi~=2019.11.28
natsort~=7.1.0
Django~=3.1.1
urllib3~=1.25.8
wcwidth~=0.2.5
simplejson~=3.17.2
jsonschema~=3.2.0
mistune~=0.8.4
marshmallow~=3.5.0
webencodings~=0.5.1
jmespath~=0.10.0
idna~=2.8
pycparser~=2.20
prettytable~=2.0.0
pyrsistent~=0.17.3
sqlparse~=0.3.1
zipp~=3.1.0